Initialize HealthCounts to non-null value.
In a stress test I was doing that pounds a circuit immediately after creation it would have thread contention on the compareAndSet, one thread would go to create it, the other(s) would go to the else and return null and cause an NPE when trying to use it.
The losing threads should get the originally initialized value set to 0.