Load configuration from hystrix-plugins.properties as well
If the file hystrix-plugins.properties exists on the class path, its
configuration will be used to setup HystrixPlugins.

This makes it possible to bundle bootstrap configuration for Hystrix
in higher-level libraries without polluting the global namespace with
a 'config.properties' file; that name should better be reserved
for top-level applications only.